+++ packages/components/lib/feedback/feedback.dart
@@ -1,1 +1,2 @@
 export 'badge/badge.dart';
+export 'separator/separator.dart';