+++ packages/components/lib/inputs/inputs.dart
@@ -1,4 +1,6 @@
 export 'checkbox/checkbox.dart';
+export 'label/label.dart';
 export 'radio/radio.dart';
 export 'switch/switch.dart';
+export 'textarea/textarea.dart';
 export 'ui_text_field.dart';